New generic parser
A new simple generic parser has been added to utils, it will
allow us to write dedicated parsers for diferent types of files.

With this class you can split a source file, a text or a buffer
into blocks. You can specify the delimiters that the parser will
detects and a group of custom tokens you want to detect with the
parser.